+++ scene/Target.svelte
@@ -12,6 +12,16 @@
 	let model = mat4.create();
 	const world_position = new Float32Array(model.buffer, 12 * 4, 3);
 
-	$: model = mat4.translate(model, $ctm, location);
+	// break `location` out into its components, so that we can
+	// skip downstream computations. TODO would be nice if there
+	// was a neater way to achieve this
+	$: x = location[0];
+	$: y = location[1];
+	$: z = location[2];
+
+	const loc = new Float32Array(3);
+	$: loc[0] = x, loc[1] = y, loc[2] = z;
+
+	$: model = mat4.translate(model, $ctm, loc);
 	$: (model, get_target(id).set(world_position));
 </script>